requirement.txt
Flask
torch
sahi
Pillow
